microservicio-usuarios/requirements.txt
boto3
PyJWT
bcrypt
pydantic
python-dotenv